Module hyper_ast::position

Positioning elements in HyperAST

Because the HyperAST is a Direct Acyclic Graph (DAG), any given sub-tree possibly has multiple global positions. The global position, path, or offset of a subtree is global/contextual information, thus it cannot be stored efficiently on subtrees of a DAG.

You can look at this module as an example of computing global metrics out of local ones.

This module specifically helps with tasks related to positioning nodes globally,

it helps maintain positional states during traversals

it helps convert between positional representations

topological

path (list of offsets)

a file path, an offset and a length

with/out hidden nodes

Incremental position storing

[structural_pos]

topological index

[topological_offset] - post-order

path

offsets_and_nodes - list of offsets

collection of path

Optimisation related, sometimes necessary to have acceptable perfs.
- list of paths
- it of paths
- topo ordered list of paths
  incremental compute
- reversed dag of paths
  mem optimization,

with hidden nodes (spaces, abtract nodes,….)

Re-exports

pub use building::CompoundPositionPreparer;

pub use offsets_and_nodes::*;

Modules

building

Declares interfaces for position builders, while offering statemachine traits to orchestrate them statisticaly.

file_and_offset

offsets

offsets_and_nodes

position_accessors

tags

Structs

ExploreStructuralPositions

PositionConverter

Scout

SpHandle

StructuralPositionStore

TypedScout

WithHyperAstPositionConverter

Traits

PrimInt

TreePath

TreePathMut

TypedTreePath

Functions

compute_position

compute_position_and_nodes

compute_position_and_nodes_with_no_spaces

compute_position_with_no_spaces

compute_range

precondition: root node do not contain a File node TODO make whole thing more specific to a path in a tree

global_pos_with_spaces

path_with_spaces

resolve_range

must be in a file

Type Definitions

Position

StructuralPosition