Struct hyper_ast::store::nodes::legion::HashedNodeRef

#[repr(transparent)]
pub struct HashedNodeRef<'a, T = NodeIdentifier>(_, _);

Implementations

impl<'a, Id: TypedNodeId<IdN = NodeIdentifier>> HashedNodeRef<'a, Id>

pub fn get_bytes_len(&self, _p_indent_len: u32) -> u32where Id::Ty: 'static + TypeTrait + Send + Sync + Debug,

impl<'a, Id: 'static + TypedNodeId<IdN = NodeIdentifier>> HashedNodeRef<'a, Id>

pub fn try_get_bytes_len(&self, _p_indent_len: u32) -> Option<u32>where Id::Ty: HyperType + Copy + Send + Sync,

pub fn is_directory(&self) -> boolwhere Id: HyperType + Copy + Send + Sync,

impl<'a, T> HashedNodeRef<'a, T>

pub fn archetype(&self) -> &Archetype

Returns the entity’s archetype.

pub fn location(&self) -> EntityLocation

Returns the entity’s location.

pub fn into_component<C: Component>(self) -> Result<&'a C, ComponentError>

Returns a reference to one of the entity’s components.

pub unsafe fn into_component_unchecked<C: Component>( self ) -> Result<&'a mut C, ComponentError>

Returns a mutable reference to one of the entity’s components.

Safety

This function bypasses static borrow checking. The caller must ensure that the component reference will not be mutably aliased.

pub fn get_component<C: Component>(&self) -> Result<&C, ComponentError>

Returns a reference to one of the entity’s components.

pub unsafe fn get_component_unchecked<C: Component>( &self ) -> Result<&mut C, ComponentError>

Returns a mutable reference to one of the entity’s components.

Safety

This function bypasses static borrow checking. The caller must ensure that the component reference will not be mutably aliased.

pub fn get_child_by_name( &self, name: &<HashedNodeRef<'a, T> as Labeled>::Label ) -> Option<<HashedNodeRef<'a, T> as Stored>::TreeId>

pub fn get_child_idx_by_name( &self, name: &<HashedNodeRef<'a, T> as Labeled>::Label ) -> Option<<HashedNodeRef<'a, T> as WithChildren>::ChildIdx>

pub fn try_get_children_name( &self ) -> Option<&[<HashedNodeRef<'a, T> as Labeled>::Label]>

impl<'a, Id: TypedNodeId<IdN = NodeIdentifier>> HashedNodeRef<'a, Id>where Id::Ty: 'static + Sync + Send + TypeTrait,

pub fn into_compressed_node( &self ) -> Result<CompressedNode<Entity, LabelIdentifier, Id::Ty>, ComponentError>

impl<'a, T> HashedNodeRef<'a, T>

pub fn size_no_spaces(&self) -> usize

impl<'a, T> HashedNodeRef<'a, T>

pub fn cs( &self ) -> Result<&<Self as WithChildren>::Children<'_>, ComponentError>

pub fn no_spaces( &self ) -> Result<&<Self as WithChildren>::Children<'_>, ComponentError>